requirements.txt
APScheduler==3.6.3
certifi==2019.11.28
chardet==3.0.4
Click==7.0
dash==1.9.0
dash-auth==1.3.2
dash-bootstrap-components==0.8.3
dash-core-components==1.8.0
dash-html-components==1.0.2
dash-renderer==1.2.4
dash-table==4.6.0
Flask==1.1.1
Flask-Compress==1.4.0
Flask-SeaSurf==0.2.2
future==0.18.2
gspread==3.2.0
gunicorn==20.0.4
httplib2==0.17.0
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
numpy==1.18.1
oauth2client==4.1.3
pandas==1.0.1
plotly==4.5.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
retrying==1.3.3
rsa==4.0
six==1.14.0
tzlocal==2.0.0
ua-parser==0.9.0
urllib3==1.25.8
Werkzeug==1.0.0
